Unify AdminEndpointResolver with Executer
