+++ layers/GraphQLAPIForWP/plugins/graphql-api-for-wp/src/Services/EndpointExecuters/AbstractClientEndpointExecuter.php
@@ -9,7 +9,7 @@ use GraphQLAPI\GraphQLAPI\Services\CustomPostTypes\GraphQLEndpointCustomPostType
 use GraphQLAPI\GraphQLAPI\Services\EndpointAnnotators\ClientEndpointAnnotatorInterface;
 use GraphQLByPoP\GraphQLClientsForWP\Clients\AbstractClient;
 
-abstract class AbstractClientEndpointExecuter extends AbstractEndpointExecuter
+abstract class AbstractClientEndpointExecuter extends AbstractCPTEndpointExecuter
 {
     private ?GraphQLCustomEndpointCustomPostType $graphQLCustomEndpointCustomPostType = null;
