+++ layers/GraphQLAPIForWP/plugins/graphql-api-for-wp/src/Services/EndpointExecuters/AbstractEndpointExecuter.php
@@ -4,9 +4,7 @@ declare(strict_types=1);
 
 namespace GraphQLAPI\GraphQLAPI\Services\EndpointExecuters;
 
-use GraphQLAPI\GraphQLAPI\Constants\RequestParams;
 use GraphQLAPI\GraphQLAPI\Registries\ModuleRegistryInterface;
-use GraphQLAPI\GraphQLAPI\Services\CustomPostTypes\GraphQLEndpointCustomPostTypeInterface;
 use PoP\BasicService\BasicServiceTrait;
 
 abstract class AbstractEndpointExecuter implements EndpointExecuterInterface
@@ -44,32 +42,6 @@ abstract class AbstractEndpointExecuter implements EndpointExecuterInterface
             return false;
         }
 
-        // Check we're loading the corresponding CPT
-        $customPostType = $this->getCustomPostType();
-        if (!\is_singular($customPostType->getCustomPostType())) {
-            return false;
-        }
-
-        // Check the endpoint is not disabled
-        global $post;
-        if (!$customPostType->isEndpointEnabled($post)) {
-            return false;
-        }
-
         return true;
     }
-
-    abstract protected function getCustomPostType(): GraphQLEndpointCustomPostTypeInterface;
-
-    /**
-     * Check the expected ?view=... is requested.
-     */
-    protected function isClientRequested(): bool
-    {
-        // Use `''` instead of `null` so that the query resolution
-        // works either without param or empty (?view=)
-        return ($_REQUEST[RequestParams::VIEW] ?? '') === $this->getView();
-    }
-
-    abstract protected function getView(): string;
 }
