+++ layers/GraphQLAPIForWP/plugins/graphql-api-for-wp/src/Services/EndpointResolvers/AbstractEndpointResolver.php
@@ -4,14 +4,12 @@ declare(strict_types=1);
 
 namespace GraphQLAPI\GraphQLAPI\Services\EndpointResolvers;
 
+use GraphQLAPI\GraphQLAPI\Services\EndpointExecuters\AbstractEndpointExecuter;
+use GraphQLAPI\GraphQLAPI\Services\EndpointExecuters\GraphQLEndpointResolverInterface;
 use GraphQLAPI\GraphQLAPI\Services\Helpers\EndpointHelpers;
-use PoP\BasicService\BasicServiceTrait;
-use PoP\Root\Services\AbstractAutomaticallyInstantiatedService;
 
-abstract class AbstractEndpointResolver extends AbstractAutomaticallyInstantiatedService implements EndpointResolverInterface
+abstract class AbstractEndpointResolver extends AbstractEndpointExecuter implements GraphQLEndpointResolverInterface
 {
-    use BasicServiceTrait;
-
     private ?EndpointHelpers $endpointHelpers = null;
 
     final public function setEndpointHelpers(EndpointHelpers $endpointHelpers): void
@@ -22,16 +20,4 @@ abstract class AbstractEndpointResolver extends AbstractAutomaticallyInstantiate
     {
         return $this->endpointHelpers ??= $this->instanceManager->getInstance(EndpointHelpers::class);
     }
-
-    public function initialize(): void
-    {
-        /**
-         * If executing the GraphQL query, resolve it
-         */
-        if ($this->isGraphQLQueryExecution()) {
-            $this->resolveGraphQLQuery();
-        }
-    }
-
-    abstract protected function resolveGraphQLQuery(): void;
 }
