--- layers/GraphQLAPIForWP/plugins/graphql-api-for-wp/src/Services/EndpointResolvers/EndpointResolverInterface.php
@@ -1,16 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace GraphQLAPI\GraphQLAPI\Services\EndpointResolvers;
-
-use GraphQLAPI\GraphQLAPI\Services\EndpointExecuters\GraphQLEndpointResolverInterface;
-
-interface EndpointResolverInterface extends GraphQLEndpointResolverInterface
-{
-    /**
-     * Indicate if it is executing a GraphQL query
-     * against the endpoint
-     */
-    public function isGraphQLQueryExecution(): bool;
-}
